Implemented Authentication

+++ Controllers/UserController.cs
@@ -1,8 +1,11 @@
 namespace learn_asp_clean_structure.Controllers;
 
+using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
 using learn_asp_clean_structure.DTOs;
 using learn_asp_clean_structure.Services;
+using Microsoft.AspNetCore.Authorization;
+using System.IdentityModel.Tokens.Jwt;
 
 [ApiController]
 [Route("api/[controller]")]
@@ -15,16 +18,22 @@ public class UserController : ControllerBase
         _userService = userService;
     }
 
-    [HttpGet]
-    public async Task<ActionResult<List<UserResponse>>> GetAll()
+    private bool IsCurrentUser(int id)
     {
-        var users = await _userService.GetAll();
-        return Ok(users);
+        var subClaimId = User.Claims.FirstOrDefault(c => c.Type == JwtRegisteredClaimNames.Sub)?.Value;
+
+        return subClaimId == id.ToString();
     }
 
     [HttpGet("{id}")]
+    [Authorize]
     public async Task<ActionResult<UserResponse>> GetById(int id)
     {
+        if (!IsCurrentUser(id))
+        {
+            return Forbid();
+        }
+
         var user = await _userService.GetById(id);
         if (user == null)
         {
@@ -43,8 +52,14 @@ public class UserController : ControllerBase
     }
 
     [HttpPut("{id}")]
+    [Authorize]
     public async Task<ActionResult<UserResponse>> Update(int id, UserUpdateRequest updatedUser)
     {
+        if (!IsCurrentUser(id))
+        {
+            return Forbid();
+        }
+
         var user = await _userService.Update(id, updatedUser);
 
         if (user == null)
@@ -56,8 +71,14 @@ public class UserController : ControllerBase
     }
 
     [HttpDelete("{id}")]
+    [Authorize]
     public async Task<ActionResult> Delete(int id)
     {
+        if (!IsCurrentUser(id))
+        {
+            return Forbid();
+        }
+
         var isDeleted = await _userService.Delete(id);
 
         if (!isDeleted)
